+++ nodedb-cluster/src/distributed_array/opcodes.rs
@@ -5,22 +5,22 @@
 //! handler code can refer to them without importing the full enum.
 
 /// Coordinator → shard: execute a coord-range slice.
-pub const ARRAY_SHARD_SLICE_REQ: u16 = 80;
+pub const ARRAY_SHARD_SLICE_REQ: u32 = 80;
 /// Shard → coordinator: slice result rows.
-pub const ARRAY_SHARD_SLICE_RESP: u16 = 81;
+pub const ARRAY_SHARD_SLICE_RESP: u32 = 81;
 /// Coordinator → shard: compute a partial aggregate.
-pub const ARRAY_SHARD_AGG_REQ: u16 = 82;
+pub const ARRAY_SHARD_AGG_REQ: u32 = 82;
 /// Shard → coordinator: partial aggregate result.
-pub const ARRAY_SHARD_AGG_RESP: u16 = 83;
+pub const ARRAY_SHARD_AGG_RESP: u32 = 83;
 /// Coordinator → shard: write cell batch.
-pub const ARRAY_SHARD_PUT_REQ: u16 = 84;
+pub const ARRAY_SHARD_PUT_REQ: u32 = 84;
 /// Shard → coordinator: put acknowledgement.
-pub const ARRAY_SHARD_PUT_RESP: u16 = 85;
+pub const ARRAY_SHARD_PUT_RESP: u32 = 85;
 /// Coordinator → shard: delete cells by exact coords.
-pub const ARRAY_SHARD_DELETE_REQ: u16 = 86;
+pub const ARRAY_SHARD_DELETE_REQ: u32 = 86;
 /// Shard → coordinator: delete acknowledgement.
-pub const ARRAY_SHARD_DELETE_RESP: u16 = 87;
+pub const ARRAY_SHARD_DELETE_RESP: u32 = 87;
 /// Coordinator → shard: surrogate bitmap scan.
-pub const ARRAY_SHARD_SURROGATE_BITMAP_REQ: u16 = 88;
+pub const ARRAY_SHARD_SURROGATE_BITMAP_REQ: u32 = 88;
 /// Shard → coordinator: surrogate bitmap result.
-pub const ARRAY_SHARD_SURROGATE_BITMAP_RESP: u16 = 89;
+pub const ARRAY_SHARD_SURROGATE_BITMAP_RESP: u32 = 89;
